Revert "AB#117017_patch_1 reinstate "grantAccess" logic"
This reverts commit 245fdd992e51d9a8c6fe7e9c3fdbf4ca85918848.
