--- src/shared/util.js
@@ -1,97 +1,0 @@
-/*
-  utils required by both the CLI and the test suite
-*/
-
-
-export function normalizeUrlParts(host, path) {
-  return {
-    host: host.trim().replace(/\/+$/, ''),
-    path: path.trim().replace(/^\/+/, '')
-  }
-}
-
-export function buildTestUrl(host, path) {
-  const { host: normalizedHost, path: normalizedPath } = normalizeUrlParts(host, path)
-  return `${normalizedHost}/${normalizedPath}`
-}
-
-/**
- * Visit `toolUrl` and complete the grant-access flow if the tool requests it.
- * Navigates the page into the LTI tool, waits for loading to finish and
- * resolves whether the tool requires an explicit grant. If required, the
- * `grantAccess` helper is used to complete the flow.
- *
- * @param {import('@playwright/test').Page} page - Playwright page instance
- * @param {import('@playwright/test').BrowserContext} context - Playwright browser context
- * @param {string} toolUrl - URL of the LTI tool to visit
- * @returns {Promise<void>}
- */
-export const grantAccessIfNeeded = async (page, context, toolUrl) => {
-  await page.goto(toolUrl)
-  const ltiToolFrame = getLtiIFrame(page)
-
-  // wait for tool-support loading page
-  await ltiToolFrame.getByText('Loading...').waitFor({
-    state: 'detached',
-    timeout: 5000,
-    strict: false
-  })
-
-  const needsGrantAccess = await Promise.race([
-    ltiToolFrame.getByText('Please Grant Access').waitFor()
-      .then(() => { return true }),
-    waitForNoSpinners(ltiToolFrame, 3000)
-      .then(() => { return false })
-  ])
-
-  if (needsGrantAccess) {
-    await grantAccess(context, ltiToolFrame)
-  }
-}
-
-/**
- * Complete the grant access flow by clicking the authorise button in the
- * popup page. Intended to be used by `grantAccessIfNeeded`.
- *
- * @param {import('@playwright/test').BrowserContext} context - Playwright browser context
- * @param {import('@playwright/test').FrameLocator} frameLocator - Locator for the LTI frame
- * @returns {Promise<void>}
- */
-const grantAccess = async (context, frameLocator) => {
-  const button = await frameLocator.getByRole('button')
-  const [newPage] = await Promise.all([
-    context.waitForEvent('page'),
-    button.click()
-  ])
-
-  const submit = await newPage.getByRole('button', { name: /Authori[sz]e/ })
-  await submit.click()
-  const close = await newPage.getByText('Close', { exact: true })
-  await close.click()
-}
-
-
-/**
- * Return the frame locator for the LTI launch iframe.
- * @param {import('@playwright/test').Page} page - Playwright page
- * @returns {import('@playwright/test').FrameLocator}
- */
-export const getLtiIFrame = (page) => {
-  return page.frameLocator('iframe[data-lti-launch="true"]')
-}
-
-/**
- * Wait for any `.view-spinner` elements inside the supplied frame locator to
- * disappear. Optionally provide an initial delay before checking.
- *
- * @param {import('@playwright/test').FrameLocator} frameLocator - Frame locator to query
- * @param {number} [initialDelay=1000] - milliseconds to wait before starting checks
- * @returns {Promise<void>}
- */
-export const waitForNoSpinners = async (frameLocator, initialDelay = 1000) => {
-  await new Promise(r => setTimeout(r, initialDelay))
-  await frameLocator.locator('.view-spinner').waitFor({
-    state: 'detached',
-    timeout: 10000
-  })
-}
